refactor(correlation): consolidate correlation models into valuations crate
- Removed the standalone `finstack-correlation` crate and integrated its functionality into the `finstack-valuations` crate, specifically under the `correlation` module.
- Updated documentation to reflect the new structure, clarifying that credit correlation infrastructure (copulas, factor models, stochastic recovery) is now part of `finstack_valuations::correlation`.
- Adjusted feature flags in `Cargo.toml` to maintain backward compatibility, retaining a deprecated `correlation` feature that points to the new location.
- Enhanced the `README.md` and other documentation files to guide users on the transition and usage of the new correlation models.

These changes streamline the codebase by centralizing correlation-related functionality, improving maintainability and user experience.

+++ finstack/valuations/tests/instruments/swaption/invariants.rs
@@ -287,7 +287,7 @@ proptest! {
         let as_of = date!(2024 - 01 - 01);
         let expiry = date!(2025 - 01 - 01);
         let swap_start = expiry;
-        let swap_end = swap_start.add_months((tenor_years as i32) * 12);
+        let swap_end = swap_start.add_months(tenor_years * 12);
 
         let payer = create_standard_payer_swaption(expiry, swap_start, swap_end, strike);
         let receiver = create_standard_receiver_swaption(expiry, swap_start, swap_end, strike);
